Simplify LineNumber constructors and fix oneshot results
