+++ core/src/main/java/org/jruby/ir/instructions/LineNumberInstr.java
@@ -1,5 +1,6 @@
 package org.jruby.ir.instructions;
 
+import org.jruby.ext.coverage.CoverageData;
 import org.jruby.ir.IRVisitor;
 import org.jruby.ir.Operation;
 import org.jruby.ir.persistence.IRReaderDecoder;
@@ -9,19 +10,35 @@ import org.jruby.ir.transformations.inlining.InlineCloneInfo;
 
 public class LineNumberInstr extends NoOperandInstr implements FixedArityInstr {
     public final int lineNumber;
+
+    /**
+     * Whether to update coverage data when executing this instruction.
+     *
+     * Note this is non-final so that `oneshot_lines` mode can disable coverage.
+     */
     public boolean coverage;
+
+    /**
+     * Whether to disable coverage after the first execution.
+     *
+     * This allows lightweight "covered or not" coverage that only updates once per line.
+     */
     public final boolean oneshot;
 
-    public LineNumberInstr(int lineNumber, boolean coverage, boolean oneshot) {
+    public LineNumberInstr(int lineNumber, int coverageMode) {
         super(Operation.LINE_NUM);
 
         this.lineNumber = lineNumber;
-        this.coverage = coverage;
-        this.oneshot = oneshot;
+        this.coverage = coverageMode != CoverageData.NONE;
+        this.oneshot = (coverageMode & CoverageData.ONESHOT_LINES) != 0;
     }
 
     public LineNumberInstr(int lineNumber) {
-        this(lineNumber, false, false);
+        super(Operation.LINE_NUM);
+
+        this.lineNumber = lineNumber;
+        this.coverage = false;
+        this.oneshot = false;
     }
 
     public int getLineNumber() {
