+++ core/src/main/java/org/jruby/runtime/profile/builtin/MethodData.java
@@ -26,6 +26,8 @@
 
 package org.jruby.runtime.profile.builtin;
 
+import org.jruby.util.collections.IntList;
+
 import java.util.ArrayList;
 
 class MethodData extends InvocationSet {
@@ -35,36 +37,6 @@ class MethodData extends InvocationSet {
     MethodData(int serial) {
         super(new ArrayList<Invocation>());
         this.serialNumber = serial;
-    }
-
-    private static class IntList {
-
-        private int[] ints = new int[10];
-        private int size;
-
-        public void add(int i) {
-            if (size == ints.length) {
-                int[] newInts = new int[(int) (ints.length * 1.5 + 1)];
-                System.arraycopy(ints, 0, newInts, 0, ints.length);
-                ints = newInts;
-            }
-            ints[size++] = i;
-        }
-
-        public boolean contains(int i) {
-            for (int j = 0; j < size; j++) {
-                if (ints[j] == i) {
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        public int[] toIntArray() {
-            int[] newInts = new int[size];
-            System.arraycopy(ints, 0, newInts, 0, size);
-            return newInts;
-        }
     }
 
     public int[] parents() {
